fix data type (#10)
fix data type (#10)

+++ src/main/java/com/github/hcsp/datatype/IntegerOverflow.java
@@ -4,8 +4,8 @@ package com.github.hcsp.datatype;
 // Fix this method to make it output "i=3000000000"
 public class IntegerOverflow {
     public static void main(String[] args) {
-        int 十亿 = 10_0000_0000;
-        int i = 0;
+        long 十亿 = 10_0000_0000;
+        long i = 0;
 
         i = i + 十亿;
         System.out.println("i=" + i);
